requirements.txt
pdfplumber
pdfminer.six
PyPDF2
pytesseract
pdf2image
requests